requirements.txt
entsoe-py==0.6.9
pandas>=2.1
numpy>=1.25
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
plotly>=5.22
streamlit>=1.36
pyyaml>=6.0
python-dotenv>=1.0
statsmodels>=0.14
holidays>=0.50
pyarrow>=17.0.0
requests>=2.32
